Update README.md
With this change, I have added language options to the README file. Users can now choose between languages at the top of the README file. This enhances the accessibility of the project for users in different languages and allows readers to access the documentation in their preferred language. The language options are neatly placed at the top of the README file and direct users to the documentation in the relevant language. This improves the user experience of the project and helps it reach a wider audience.